requirements.txt
streamlit
pymupdf
openai
python-dotenv
pillow
pandas
